Allow operator to assume ownership of existing IP pools
