+++ api/v1/installation_types.go
@@ -18,8 +18,6 @@ package v1
 
 import (
 	"fmt"
-	"net"
-	"net/netip"
 	"strings"
 
 	pcv1 "github.com/tigera/operator/pkg/apis/crd.projectcalico.org/v1"
@@ -596,6 +594,9 @@ func (nt NATOutgoingType) String() string {
 const NodeSelectorDefault string = "all()"
 
 type IPPool struct {
+	// Name is the name of the IP pool. If omitted, this will be generated.
+	Name string `json:"name,omitempty"`
+
 	// CIDR contains the address range for the IP Pool in classless inter-domain routing format.
 	CIDR string `json:"cidr"`
 
@@ -630,44 +631,11 @@ type IPPool struct {
 	DisableBGPExport *bool `json:"disableBGPExport,omitempty"`
 }
 
-// cidrToName returns a valid Kubernetes resource name given a CIDR. Kubernetes names must be valid DNS
-// names. We do the following:
-// - Expand the CIDR so that we get consistent results and remove IPv6 shorthand "::".
-// - Replace any slashes with dashes.
-// - Replace any : with dots.
-func cidrToName(cidr string) (string, error) {
-	// First, canonicalize the CIDR. e.g., 192.168.0.1/24 -> 192.168.0.0/24.
-	_, nw, err := net.ParseCIDR(cidr)
-	if err != nil {
-		return "", err
-	}
-
-	// Parse the CIDR and expand it to its full form.
-	// e.g., fe80::/64 -> fe80:0000:0000:0000:0000:0000:0000:0000/64
-	pre, err := netip.ParsePrefix(nw.String())
-	if err != nil {
-		return "", err
-	}
-	name := pre.Addr().StringExpanded()
-
-	// Replace invalid characters.
-	// e.g., fe80:0000:0000:0000:0000:0000:0000:0000/64 -> fe80.0000.0000.0000.0000.0000.0000.0000-64
-	name = strings.ReplaceAll(name, ":", ".")
-	name += fmt.Sprintf("-%d", pre.Bits())
-
-	return name, nil
-}
-
 // ToProjectCalicoV1 converts an IPPool to a crd.projectcalico.org/v1 IPPool resource.
 func (p *IPPool) ToProjectCalicoV1() (*pcv1.IPPool, error) {
-	name, err := cidrToName(p.CIDR)
-	if err != nil {
-		return nil, err
-	}
-
 	pool := pcv1.IPPool{
 		TypeMeta:   metav1.TypeMeta{Kind: "IPPool", APIVersion: "crd.projectcalico.org/v1"},
-		ObjectMeta: metav1.ObjectMeta{Name: name},
+		ObjectMeta: metav1.ObjectMeta{Name: p.Name},
 		Spec:       pcv1.IPPoolSpec{CIDR: p.CIDR},
 	}
